fix(staking): widen RewardSchedule.lastUpdate to uint64 (audit)
Auditor flagged the uint32 timestamp watermark as inadequate. lastUpdate
stored uint32(block.timestamp), which truncates at 2**32 seconds (~year
2106) and was narrower than the uint64 `start` it is compared against in
_vestStream (last >= start, last - start) -- past the boundary the wrapped
watermark makes vested-so-far collapse to 0 and the schedule re-vests from
the beginning (over-distribution).

Repack RewardSchedule so lastUpdate is uint64 at no extra slot cost: move
nextDuration up beside duration to fill slot 0, freeing a 64-bit gap in
slot 1 for the widened lastUpdate. Still exactly two fully-packed slots.

- IVanaPoolEntity.RewardSchedule: reorder + lastUpdate uint32 -> uint64.
- lastUpdate write cast uint32 -> uint64.
- Reorder positional RewardSchedule literals in tests (lastUpdate now last)
  and widen the casts; VestStream named literals widened.
- New VestStream regression vesting past the 2**32 boundary. Suite 144/144.

Durations stay uint32 (~136 yr span, adequate, and what keeps the 2-slot
packing).

+++ contracts/vanaStaking/vanaPoolEntity/VanaPoolEntityImplementation.sol
@@ -1120,7 +1120,7 @@ contract VanaPoolEntityImplementation is
         // further queued entry it already wrote this same value (a redundant
         // but harmless SSTORE); when it promoted an empty slot it short-circuited
         // on the scheduledValue == 0 guard without writing, so this is required.
-        schedule.lastUpdate = uint32(block.timestamp);
+        schedule.lastUpdate = uint64(block.timestamp);
     }
 
     /**
